templates: keep device IDs through out the state updates
If the user issues a new command for a device that was mentioned
previously in the context, assume they want to refer to the same
device, unless they specify a new name or "all devices". This
allows to slot-fill commands for devices with multiple instances,
because the model learns to propagate the device ID across turn
and we don't lose information.